+++ src/mainReadout.cxx
@@ -603,6 +603,9 @@ int Readout::configure(const boost::property_tree::ptree& properties)
     return -1;
   }
 
+  // TODO
+  // save config somewhere ?
+
   // apply provided occ properties over loaded configuration
   // with function to overwrtie configuration tree t1 with (selected) content of t2
   auto mergeConfig = [&](boost::property_tree::ptree& t1, const boost::property_tree::ptree& t2) {
@@ -618,7 +621,7 @@ int Readout::configure(const boost::property_tree::ptree& properties)
           // check for a consumer with same fairmq channel
           for (auto kName : ConfigFileBrowser(&cfg, "consumer-")) {
             std::string cfgType;
-            cfgType = cfg.getValue<std::string>(kName + ".consumerType");
+            cfg.getOptionalValue<std::string>(kName + ".consumerType", cfgType);
             if (cfgType == "FairMQChannel") {
               std::string cfgChannelName;
               cfg.getOptionalValue<std::string>(kName + ".fmq-name", cfgChannelName);
@@ -659,6 +662,43 @@ int Readout::configure(const boost::property_tree::ptree& properties)
     }
   };
   mergeConfig(cfg.get(), properties);
+
+  // merge default trees
+  const char *defaultTag = "-*";
+  // if a section ends with above string, its parameters are copied to all matching section names, when these parameters are not already defined there
+  // append configuration tree t1 with content of t2 (existing leaves in t1 not overwritten by same ones from t2)
+  auto appendConfig = [&](boost::property_tree::ptree& t1, boost::property_tree::ptree& t2) {
+    for (boost::property_tree::ptree::iterator it = t2.begin(); it != t2.end();++it) {
+      if (!t1.get_child_optional(it->first)) {
+        //printf("set %s = %s\n", it->first.c_str(),it->second.data().c_str());
+        t1.put_child( it->first, it->second );
+      }
+    }
+  };
+  // function returns true when end of string matches tag
+  auto isEndOfStringMatching = [&](const std::string &s, const char* tag) {
+    if (s.length()<=strlen(tag)) return false;
+    if (s.substr(s.length()-strlen(tag)) != tag) return false;
+    return true;
+  };
+  for (boost::property_tree::ptree::iterator pos = cfg.get().begin(); pos != cfg.get().end();) {    
+    const std::string section = pos->first;
+    if (!isEndOfStringMatching(section, defaultTag)) {
+      ++pos;
+      continue;
+    }
+    auto smatch = section.substr(0,section.length()-strlen(defaultTag));
+    for (boost::property_tree::ptree::iterator pos2 = cfg.get().begin(); pos2 != cfg.get().end();++pos2) {
+      if (pos2 == pos) continue; // do not self-overwrite
+      const std::string section2 = pos2->first;
+      if (section2.compare(0,smatch.length(),smatch)) continue; // mismatch
+      if (isEndOfStringMatching(section2, defaultTag)) continue; // do not overwrite other defaults sections
+      theLog.log(LogInfoDevel_(3002), "Updating configuration section [%s] with defaults from [%s]", pos2->first.c_str(), pos->first.c_str());
+      appendConfig(pos2->second,pos->second);
+    }
+    // delete section with defaults, to avoid it is used further
+    pos = cfg.get().erase(pos);
+  }
 
   // extract optional configuration parameters
   // configuration parameter: | readout | customCommands | string | | List of key=value pairs defining some custom shell commands to be executed at before/after state change commands. |
@@ -846,11 +886,8 @@ int Readout::configure(const boost::property_tree::ptree& properties)
   for (auto kName : ConfigFileBrowser(&cfg, "bank-")) {
     // skip disabled
     int enabled = 1;
-    try {
-      // configuration parameter: | bank-* | enabled | int | 1 | Enable (1) or disable (0) the memory bank. |
-      enabled = cfg.getValue<int>(kName + ".enabled");
-    } catch (...) {
-    }
+    // configuration parameter: | bank-* | enabled | int | 1 | Enable (1) or disable (0) the memory bank. |
+    cfg.getOptionalValue<int>(kName + ".enabled", enabled);
     if (!enabled) {
       continue;
     }
@@ -868,13 +905,9 @@ int Readout::configure(const boost::property_tree::ptree& properties)
     // bank type
     // configuration parameter: | bank-* | type | string| | Support used to allocate memory. Possible values: malloc, MemoryMappedFile. |
     std::string cfgType = "";
-    try {
-      cfgType = cfg.getValue<std::string>(kName + ".type");
-    } catch (...) {
-      theLog.log(LogErrorSupport_(3100), "Skipping memory bank %s:  no type specified", kName.c_str());
-      continue;
-    }
+    cfg.getOptionalValue<std::string>(kName + ".type", cfgType);
     if (cfgType.length() == 0) {
+      theLog.log(LogErrorSupport_(3100), "Skipping memory bank %s:  no type specified", kName.c_str());
       continue;
     }
 
@@ -934,11 +967,8 @@ int Readout::configure(const boost::property_tree::ptree& properties)
 
     // skip disabled
     int enabled = 1;
-    try {
-      // configuration parameter: | consumer-* | enabled | int | 1 | Enable (value=1) or disable (value=0) the consumer. |
-      enabled = cfg.getValue<int>(kName + ".enabled");
-    } catch (...) {
-    }
+    // configuration parameter: | consumer-* | enabled | int | 1 | Enable (value=1) or disable (value=0) the consumer. |
+    enabled = cfg.getOptionalValue<int>(kName + ".enabled", enabled);
     if (!enabled) {
       continue;
     }
@@ -957,7 +987,11 @@ int Readout::configure(const boost::property_tree::ptree& properties)
     try {
       // configuration parameter: | consumer-* | consumerType | string |  | The type of consumer to be instanciated. One of:stats, FairMQDevice, DataSampling, FairMQChannel, fileRecorder, checker, processor, tcp. |
       std::string cfgType = "";
-      cfgType = cfg.getValue<std::string>(kName + ".consumerType");
+      cfg.getOptionalValue<std::string>(kName + ".consumerType", cfgType);
+      if (cfgType.length() == 0) {
+        theLog.log(LogErrorSupport_(3100), "Skipping consumer %s:  no type specified", kName.c_str());
+        continue;
+      }
       theLog.log(LogInfoDevel, "Configuring consumer %s: %s", kName.c_str(), cfgType.c_str());
 
       #ifdef WITH_NUMA
@@ -1090,7 +1124,12 @@ int Readout::configure(const boost::property_tree::ptree& properties)
 
     // configuration parameter: | equipment-* | equipmentType | string |  | The type of equipment to be instanciated. One of: dummy, rorc, cruEmulator |
     std::string cfgEquipmentType = "";
-    cfgEquipmentType = cfg.getValue<std::string>(kName + ".equipmentType");
+    cfg.getOptionalValue<std::string>(kName + ".equipmentType", cfgEquipmentType);
+    if (cfgEquipmentType.length() == 0) {
+      theLog.log(LogErrorSupport_(3100), "Skipping equipment %s:  no type specified", kName.c_str());
+      continue;
+    }
+
     theLog.log(LogInfoDevel, "Configuring equipment %s: %s", kName.c_str(), cfgEquipmentType.c_str());
 
     #ifdef WITH_NUMA